Extract NodeMap
